Add dates to unclosed parts and missing tests; format() df before printing
Added some dates to the triple-negative-but-open and missing-test reports -
makes it easier to sort and identify the most egregious cases.

In dfprint, I now format() the data.frames before passing them to xtable.
print.xtable() gets crazy when you pass Date or POSIXct variables to it;
format() converts all of the variables to character before they even see
xtable().