Let xtable() handle NAs instead of format()
By default, format() encodes NAs as "NA" - meaning the report was suddenly
full of "NA" instead of "-". By setting na.encode = FALSE, any NAs in
character vectors are passed to xtable as NA, not "NA" - so that xtable
can then convert them to "-".

NAs in logical vectors are still encoded as NA, according to the docs.